+++ Source/SysPSP/HLEAudio/AudioPluginPSP.cpp
@@ -110,7 +110,7 @@ void	CAudioPluginPsp::StopEmulation()
 void	CAudioPluginPsp::DacrateChanged( int SystemType )
 {
 //	printf( "DacrateChanged( %s )\n", (SystemType == ST_NTSC) ? "NTSC" : "PAL" );
-	u32 type {(SystemType == ST_NTSC) ? VI_NTSC_CLOCK : VI_PAL_CLOCK};
+	u32 type {(u32)((SystemType == ST_NTSC) ? VI_NTSC_CLOCK : VI_PAL_CLOCK)};
 	u32 dacrate {Memory_AI_GetRegister(AI_DACRATE_REG)};
 	u32	frequency {type / (dacrate + 1)};
