+++ Source/SysPSP/HLEGraphics/GraphicsPluginPSP.cpp
@@ -196,8 +196,8 @@ void CGraphicsPluginImpl::UpdateScreen()
 
 		//Calc sync rates for audio and game speed //Corn
 		const f32 inv_Fsync {1.0f / Fsync};
-		gSoundSync = {44100.0f * inv_Fsync};
-		gVISyncRate = {1500.0f * inv_Fsync};
+		gSoundSync = {(u32)(44100.0f * inv_Fsync)};
+		gVISyncRate = {(u32)(1500.0f * inv_Fsync)};
 		if( gVISyncRate > 4000 ) gVISyncRate = 4000;
 		else if ( gVISyncRate < 1500 ) gVISyncRate = 1500;
